Moved the NEVPT2 amplitudes code into a separate file to be consistent with the MR-ADC implementation
